add clamav 0.9.73
It is based on the buildtool setup contributed by Alejandro Dguez.
I've updated the source and reworked the buildtool files.

Currently we have no package that uses clamav, but I can think about
a proxy with virus-checking. In any case the work shouldn't be lost :)

Note: clamav is not a typical application one wants to use on a common
LEAF box. It needs somewhat a lot of RAM (about 100MB to start it)
and at least 120MB storage (either RAM or a CF/HD/USB device).