update the config patch for the versatile platform
*Please* review!
Before the the update we used patch based an old config (3.2.13?), which failed to apply.
I've also added the config as bering-3.2.29-versatile, cause it's a lot easier to run
"make oldconfig" with an old config in stock, instead to get an old version from git to build linux
with the patch applied...

Not shure if it cures the pb, when I set the toolchain in buildtool.local, but at least
buildtool.pl -t armv5te-unknown-linux-uclibcgnueabi build linux|kernel|etc works with it.
As I said review the config, I haven't tested the resulting kernel neither with qemu nor anything else.