add package fwlogwatch
The documentation for fwlogwtach is sparse, so the interested user is mostly on it's own.

I checked that daemon mode works, notification via email works as well, see
/etc/fwlogwatch/fwlw_notify
/etc/cron.daily/fwlogwatch needs modifications by the user to run.

The respond mode "does something" (:= add's iptable rules) , though I can't find documentation
how to control it.

The fwlogwatch_small.cgi also works - as long as shorewall.log is readable by group adm AND user sh-httpd
is in group adm. (It needs modification and testing of shorewall package to make group adm readibilty sustainable,
but that's a question for the devel-list).

As this package has been adeed by user request, it's currentyl provided AS-IS for testing and feedback.