linux: unexport CROSS_COMPILE env variable, fix update to 3.2.31
When linux source is prepared, toolchain is still not built (it's dependent on linux headers), which results in harmless error message at headers installation process. Maybe there is cleaner solution - but in any case it works.
Also fixed breakage due to kernel update (KernelVersion in buildtool.cfg still pointed to 3.2.29)