new bootmenu for vga versions with the new LEAF logo.
based on the latest chnages in master (syslinux [vesa]menu.c32 and hdt.c32 and leaf.png)
the vga versions does have new cfg's for syslinux/isolinux and pxelinux) with a graphical bootmenu.
It provides also a hardware detection utility (without pci capabilities and without memtest86), might be removed for
the final version, but as a teaser to play with right now.

I've also tested such menus on an Alix box with serial output only. Worked, but more testing is welcome.

syslinux.dpy should be deleted later for housekeeping.
Also it's questionable to enable a special Autoboot message with a Bering-uClibc version shown.
Not shure if that is a way to go, cause it won't change if users just update and do not update syslinux/isolinux.cfg...