shorewall: move logging from syslog-ng to ulogd
Successfully running for a long time on my box and solved issues with syslog-ng
reported by a user on leaf-user. I did a lot of testing with syslog-ng, and found that
it's too unrealiable; I could reprodruce the erros and found no solution.
So I've decided to switch to a solution (ulogd) that we've used earlier successfully, and
renmoved for some time cause ulogd didn't work with the start of 5.x.
Now it does again.
Anyway this is a major change, therefor I'd like to have another rc before 5.0.2 final.

Also remove shorewall(6) from lrp.conf log rotation, this is now done in
ulogds cron file.