dnsmasq : build two packages with and without DNSSEC support
dnsmasq.lrp is the known package
dnsmasqsec.lrp is the package with DNSSEC support

dnsmasqsec.lrp requires libgmp and libnettle

To enable and test DNSSEC support From dnsmasq mailinglist:
"So, adding
 conf-file=/path/to/trust-anchors.conf
 dnssec
to your config is all thats needed to get things
working. The upstream nameservers have to be DNSSEC-capable
too, of course. Many ISP nameservers aren't, but the
Google public nameservers (8.8.8.8 and 8.8.4.4) are."

Enable log-queries to show results.

Point your browser to
 paypal.com
 ietf.org
 www.dnssec-failed.org

and see how/if it works.