update kernel to 3.14.13
kernel 3.14 is the latest longtermn stable kernel, so it might be a good base
for 5.2

- PLEASE review the kernel configs!

Note: only the i486, i686 and geode kernel has been updated, x86_64 and arm*
toolchain will fail because the patches are still on 3.10.
Anyway I have up and running a 3.14.13 router (geode) for a while, so those updated seems looking fine.